requirements.txt
keras=2.14.0
matplotlib=3.8.0
numpy=1.26.2
opencv=4.6.0
pandas=2.1.4
pillow=10.0.1
plotly=5.9.0
python=3.10.13
seaborn=0.12.1
scikit-learn=1.3.0
streamlit=1.31.1
streamlit-image-select=0.6.0
tensorflow=2.14.0
tqdm=4.65.0